move getVariableContent
